refactor: split utils into files
